requirements.txt
Flask==2.0.2
pyserial==3.5
